Emulate interrupt like behavior in vsoc_usb.
With the separation of hcd and gadget drivers, the communication between
them is through shared memory and signaling is through interrupts.
There are two parts to having interrupt support
1) generation of interrupts
2) handing of interrupts.

This bug is addressing generation of interrupt like behavior when the
hcd and gadget drivers are in the same domain/kernel (which is an
intermediate step before moving them two different machines).
For the lack of proper interrupts on the emulated hcd and gadget
devices, the hcd and gadget devices will register with the shm driver
with callback routines to raise events. For example, 'hcd' will register
guest_to_hcd_kick() and 'gadget' will register hcd_to_gadget_kick()
routines with the shm driver. These routines internally schedule
tasklets (mimicking interrupt like behavior).
The events are acknowledged in the tasklet (like a top half of an
interrupt) and the actual handling of the events are done in kernel
threads.

BUG=65946438

Change-Id: Ia7db091a052543a40b8b443ad51a0b3d0375315f
(cherry picked from commit 4768b54a73465513bef3cf24ef1c5f55e96e49db)